requirements.txt
tkinter
calendar
smtplib
python-dotenv
ics
reportlab
pandas
